requirements.txt
psycopg2-binary
pgvector
neo4j
sentence-transformers
python-dotenv
pdfplumber
pypdf
